backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
MarkupSafe==3.0.2
python-docx==1.1.2
typing_extensions==4.12.2
Werkzeug==3.1.3
flask-cors==4.0.0
